[ALS-6567] Move FENCE PSAMA permissions to "on-start up"
Refactored FENCEAuthenticationService moving all Role, Privilege, and AccessRule to their related services. Refactored all unit test and classes to account for these changes. Added a new post construct to the RoleService class that will generate all potential roles, privileges, and access rules, if the fence_mapping.json is present.
