+++ pic-sure-auth-services/src/main/java/edu/harvard/hms/dbmi/avillach/auth/repository/RoleRepository.java
@@ -4,6 +4,7 @@ import edu.harvard.hms.dbmi.avillach.auth.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
@@ -19,5 +20,8 @@ public interface RoleRepository extends JpaRepository<Role, UUID> {
 
     // find all user roles where uuid is in the list of uuids
     Set<Role> findByUuidIn(Set<UUID> uuids);
+
+    // find all role names where name in set
+    List<Role> findByNameIn(Set<String> roleNames);
 
 }
