+++ pic-sure-auth-services/src/main/java/edu/harvard/hms/dbmi/avillach/auth/service/impl/authorization/AuthorizationService.java
@@ -161,5 +161,4 @@ public class AuthorizationService {
 
 
 
-
 }
