+++ pic-sure-auth-services/src/main/java/edu/harvard/hms/dbmi/avillach/auth/utils/StringUtils.java
@@ -1,0 +1,41 @@
+package edu.harvard.hms.dbmi.avillach.auth.utils;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class StringUtils {
+
+    public static String extractProject(String roleName) {
+        String projectPattern = "FENCE_(.*?)(?:_c\\d+)?$";
+        if (roleName.startsWith("MANUAL_")) {
+            projectPattern = "MANUAL_(.*?)(?:_c\\d+)?$";
+        }
+        Pattern projectRegex = Pattern.compile(projectPattern);
+        Matcher projectMatcher = projectRegex.matcher(roleName);
+        String project = "";
+        if (projectMatcher.find()) {
+            project = projectMatcher.group(1).trim();
+        } else {
+            String[] parts = roleName.split("_", 1);
+            if (parts.length > 0) {
+                project = parts[1];
+            }
+        }
+        return project;
+    }
+
+    public static String extractConsentGroup(String roleName) {
+        String consentPattern = "FENCE_.*?_c(\\d+)$";
+        if (roleName.startsWith("MANUAL_")) {
+            consentPattern = "MANUAL_.*?_c(\\d+)$";
+        }
+        Pattern consentRegex = Pattern.compile(consentPattern);
+        Matcher consentMatcher = consentRegex.matcher(roleName);
+        String consentGroup = "";
+        if (consentMatcher.find()) {
+            consentGroup = "c" + consentMatcher.group(1).trim();
+        }
+        return consentGroup;
+    }
+
+}
