+++ pic-sure-auth-services/src/test/java/edu/harvard/hms/dbmi/avillach/auth/service/PrivilegeServiceTest.java
@@ -2,6 +2,7 @@ package edu.harvard.hms.dbmi.avillach.auth.service;
 
 import edu.harvard.hms.dbmi.avillach.auth.entity.Privilege;
 import edu.harvard.hms.dbmi.avillach.auth.repository.PrivilegeRepository;
+import edu.harvard.hms.dbmi.avillach.auth.service.impl.PrivilegeService;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.InjectMocks;
