+++ pic-sure-auth-services/src/test/java/edu/harvard/hms/dbmi/avillach/auth/service/impl/AuthorizationServiceTest.java
@@ -27,6 +27,7 @@ public class AuthorizationServiceTest {
     private SecurityContext securityContext;
 
     private AuthorizationService authorizationService;
+
     private AccessRuleService accessRuleService;
 
     @Mock
@@ -37,7 +38,7 @@ public class AuthorizationServiceTest {
         MockitoAnnotations.initMocks(this);
         SecurityContextHolder.setContext(securityContext);
 
-        accessRuleService = new AccessRuleService(accessRuleRepository);
+        accessRuleService = new AccessRuleService(accessRuleRepository, "false", "false", "false", "false", "false", "false");
         authorizationService = new AuthorizationService(accessRuleService, "fence,okta,open");
     }
